requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
